+++ F1005/Areas/InsurancesandFund/Controllers/InsurancesController.cs
@@ -31,6 +31,7 @@ namespace F1005.Areas.InsurancesandFund.Controllers
         }
 
         [HttpPost]
+        [ValidateAntiForgeryToken]
         public ActionResult Withdraw(Insurances insurances)
         {
             if (Session["User"] == null)
@@ -38,33 +39,76 @@ namespace F1005.Areas.InsurancesandFund.Controllers
                 return RedirectToRoute("Default", new { Controller = "Home", Action = "Index" });
             }
 
-            Insurances olddata = db.Insurances.Find(insurances.SerialNumber);
-            olddata.Withdrawed = true;
-            db.SaveChanges();
-            insurances.CashFlow = insurances.Withdrawal;
-            insurances.PurchaseOrWithdraw = false;
-            insurances.Withdrawed = true;
-            insurances.RelateCash = true;
-            insurances.UserID = Session["User"].ToString();
-            SummaryTable ST = new SummaryTable { TradeType = "保險", TradeDate = insurances.WithdrawDate, UserName = insurances.UserID };
-            db.SummaryTable.Add(ST);
-            db.SaveChanges();
-            var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
-            insurances.STID = id;
-            db.Insurances.Add(insurances);
-            db.SaveChanges();
-
-            CashIncome CI = new CashIncome
+            if (ModelState.IsValid)
             {
-                OID = id,
-                UserName = insurances.UserID,
-                InCashType = "保險",
-                InAmount = insurances.CashFlow,
-                InDate = insurances.WithdrawDate,
-                InNote = insurances.InsuranceName+"解約金"
-            };
-            db.CashIncome.Add(CI);
-            return RedirectToAction("Index");
+                SummaryTable ST = new SummaryTable
+                {
+                    TradeType = "保險",
+                    TradeDate = insurances.WithdrawDate,
+                    UserName = Session["User"].ToString()
+                };
+
+
+
+                Insurances olddata = db.Insurances.Find(insurances.SerialNumber);
+                olddata.Withdrawed = true;
+                db.SaveChanges();
+
+                insurances.PurchaseOrWithdraw = false;
+                insurances.Withdrawed = true;
+
+                insurances.UserID = Session["User"].ToString();
+                if (insurances.RelateCash == true)
+                {
+
+                    insurances.CashFlow = insurances.Withdrawal;
+                    db.SummaryTable.Add(ST);
+                    db.SaveChanges();
+
+                    var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
+                    insurances.STID = id;
+                    db.Insurances.Add(insurances);
+                    db.SaveChanges();
+
+                    var User = db.UsersData.Where(U => U.UserName == insurances.UserID).Select(U => U).SingleOrDefault();
+                    User.InsuranceValue -= insurances.Withdrawal;
+                    User.CashValue += insurances.Withdrawal;
+                    db.SaveChanges();
+
+                    CashIncome CI = new CashIncome
+                    {
+                        OID = id,
+                        UserName = insurances.UserID,
+                        InCashType = "保險",
+                        InAmount = insurances.CashFlow,
+                        InDate = insurances.WithdrawDate,
+                        InNote = insurances.InsuranceName + "解約金"
+                    };
+
+                    db.CashIncome.Add(CI);
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                else
+                {
+                    insurances.CashFlow = 0;
+                    db.SummaryTable.Add(ST);
+                    db.SaveChanges();
+
+                    var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
+                    insurances.STID = id;
+                    db.Insurances.Add(insurances);
+                    db.SaveChanges();
+
+                    var User = db.UsersData.Where(U => U.UserName == insurances.UserID).Select(U => U).SingleOrDefault();
+                    User.InsuranceValue -= insurances.Withdrawal;
+                    User.CashValue += insurances.Withdrawal;
+                    db.SaveChanges();
+
+                    return RedirectToAction("Index");
+                }
+            }
+            return View(insurances);
         }
 
         // 保險首頁
@@ -77,8 +121,6 @@ namespace F1005.Areas.InsurancesandFund.Controllers
 
             return View(db.Insurances.ToList());
         }
-
-        
         //新增保單
         // GET: Insurances/Create
         public ActionResult Create()
@@ -90,55 +132,75 @@ namespace F1005.Areas.InsurancesandFund.Controllers
 
             return View();
         }
-
         // POST: Insurances/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "SerialNumber,UserID,InsuranceName,PurchaseDate,WithdrawDate,PaymentPerYear,PayYear,PurchaseOrWithdraw,CashFlow,Withdrawal,RelateCash")] Insurances insurances)
         {
-            if (insurances.RelateCash == true)
-            {
-                insurances.CashFlow = -insurances.PaymentPerYear * insurances.PayYear;
-            }
-            else
-            {
-                insurances.CashFlow = 0;
-            }
-            insurances.PurchaseOrWithdraw = true;
-            insurances.Withdrawed = false;
-
-            //insurances.UserID = Session["User"].ToString();
             if (ModelState.IsValid)
             {
                 insurances.UserID = Session["User"].ToString();
-                SummaryTable ST = new SummaryTable { TradeType = "保險", TradeDate = insurances.PurchaseDate, UserName = insurances.UserID };
-                db.SummaryTable.Add(ST);
-                db.SaveChanges();
-                var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
-                insurances.STID = id;
-                db.Insurances.Add(insurances);
-                db.SaveChanges();
+                insurances.PurchaseOrWithdraw = true;
+                insurances.Withdrawed = false;
 
-                CashExpense CE = new CashExpense();
-                CE.ExAmount = -insurances.CashFlow;
-                CE.UserName = insurances.UserID;
-                CE.ExDate = ST.TradeDate;
-                CE.ExCashType = ST.TradeType;
-                CE.OID = id;
-                CE.ExNote = insurances.InsuranceName + "支出";
+                SummaryTable ST = new SummaryTable
+                {
+                    TradeType = "保險",
+                    TradeDate = insurances.PurchaseDate,
+                    UserName = insurances.UserID
+                };
+                string UID = Session["User"].ToString();
+                var ExSUM = db.CashExpense.Where(C => C.UserName == UID).Sum(C => C.ExAmount);
+                var InSUM = db.CashIncome.Where(C => C.UserName == UID).Sum(C => C.InAmount);
 
-                db.CashExpense.Add(CE);
-                db.SaveChanges();
+                if (insurances.RelateCash == true&&InSUM-ExSUM> insurances.PaymentPerYear * insurances.PayYear)
+                {
+                    db.SummaryTable.Add(ST);
+                    db.SaveChanges();
+
+                    var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
+                    insurances.STID = id;
+                    insurances.CashFlow = -insurances.PaymentPerYear * insurances.PayYear;
+
+                    db.Insurances.Add(insurances);
+                    db.SaveChanges();
+
+                    CashExpense CE = new CashExpense
+                    {
+                        ExAmount = -insurances.CashFlow,
+                        UserName = insurances.UserID,
+                        ExDate = ST.TradeDate,
+                        ExCashType = ST.TradeType,
+                        OID = id,
+                        ExNote = insurances.InsuranceName + "支出"
+                    };
+                    db.CashExpense.Add(CE);
+
+                    var User = db.UsersData.Where(U => U.UserName == insurances.UserID).Select(U => U).SingleOrDefault();
+                    User.InsuranceValue += insurances.Withdrawal;
+                    User.CashValue -= insurances.PaymentPerYear * insurances.PayYear;
+                    db.SaveChanges();
+                }
+                else
+                {
+                    db.SummaryTable.Add(ST);
+                    db.SaveChanges();
+
+                    var id = db.SummaryTable.Select(c => c.STId).ToList().LastOrDefault();
+                    insurances.STID = id;
+                    insurances.CashFlow = 0;
+                    db.Insurances.Add(insurances);
+                    db.SaveChanges();
+
+                    var User = db.UsersData.Where(U => U.UserName == insurances.UserID).Select(U => U).SingleOrDefault();
+                    User.InsuranceValue += insurances.Withdrawal;
+                    db.SaveChanges();
+                }
                 return RedirectToAction("Index");
-              
             }
-
             return View(insurances);
         }
-
-     
-
-
+        //Dispose
         protected override void Dispose(bool disposing)
         {
             if (disposing)
@@ -147,23 +209,22 @@ namespace F1005.Areas.InsurancesandFund.Controllers
             }
             base.Dispose(disposing);
         }
-
         //計算IRR
         public ActionResult CalculateIRR(IRRViewModel model)
         {
             IRRCalculater calculater = new IRRCalculater();
+
             return Content(calculater.IRR(model));
         }
-
         //生成未實現圖表
         public JsonResult GetCurrentDoughnut()
         {
             string UID = Session["User"].ToString();
-            var query = db.Insurances.Where(I => I.PurchaseOrWithdraw == true && I.UserID == UID&&I.Withdrawed==false).Select(I => new
+            var query = db.Insurances.Where(I => I.PurchaseOrWithdraw == true && I.UserID == UID && I.Withdrawed == false).Select(I => new
 
             {
                 Name = I.InsuranceName,
-                Money = I.PaymentPerYear*I.PayYear
+                Money = I.PaymentPerYear * I.PayYear
             }).ToList();
             return Json(query, JsonRequestBehavior.AllowGet);
         }
@@ -176,10 +237,10 @@ namespace F1005.Areas.InsurancesandFund.Controllers
             {
                 Name = I.InsuranceName,
                 Money = I.Withdrawal
-            } ).ToList();
+            }).ToList();
             return Json(query, JsonRequestBehavior.AllowGet);
         }
-
+        //load data to 交易紀錄
         public ActionResult loadalldata()
         {
             string UID = Session["User"].ToString();
@@ -191,67 +252,76 @@ namespace F1005.Areas.InsurancesandFund.Controllers
                 InsuranceName = I.InsuranceName,
                 PurchaseDate = I.PurchaseDate.ToShortDateString(),
                 WithdrawDate = I.WithdrawDate.ToShortDateString(),
-                PaymentPerYear = I.PaymentPerYear,
+                PaymentPerYear = I.PaymentPerYear.Value.ToString("C0"),
                 PayYear = I.PayYear,
                 PurchaseOrWithdraw = I.PurchaseOrWithdraw,
-                Withdrawal = I.Withdrawal,
+                Withdrawal = I.Withdrawal.Value.ToString("C0"),
                 Withdrawed = I.Withdrawed,
                 CashFlow = I.CashFlow,
             });
             return Json(items, JsonRequestBehavior.AllowGet);
         }
-
         //Edit 
         public ActionResult EditIinsurance(Insurances insurance)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(insurance).State = EntityState.Modified;
+
+                var stdata = db.SummaryTable.Find(insurance.STID);
+                if (insurance.PurchaseOrWithdraw == true)
+                {
+                    stdata.TradeDate = insurance.PurchaseDate;
+                    var arr = db.CashExpense.Where(C => C.OID == insurance.STID).Select(C => C).SingleOrDefault();
+                    arr.ExAmount = insurance.CashFlow;
+                    arr.ExDate = insurance.PurchaseDate;
+                }
+                else
+                {
+                    stdata.TradeDate = insurance.WithdrawDate;
+                    var arr = db.CashIncome.Where(C => C.OID == insurance.STID).Select(C => C).SingleOrDefault();
+                    arr.InAmount = -insurance.CashFlow;
+                    arr.InDate = insurance.PurchaseDate;
+                }
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            var stdata = db.SummaryTable.Find(insurance.STID);
-            if (insurance.PurchaseOrWithdraw == true)
-            {
-                stdata.TradeDate = insurance.PurchaseDate;
-                var arr = db.CashExpense.Where(C => C.OID == insurance.STID).Select(C => C).ToArray();
-                arr[0].ExAmount = insurance.CashFlow;
-                arr[0].ExDate = insurance.PurchaseDate;
-            }
-            else
-            {
-                stdata.TradeDate = insurance.WithdrawDate;
-                var arr = db.CashIncome.Where(C => C.OID == insurance.STID).Select(C => C).ToArray();
-                arr[0].InAmount = -insurance.CashFlow;
-                arr[0].InDate = insurance.PurchaseDate;
-            }
-          
-
             return View(insurance);
         }
-
         //Delete 
         public ActionResult DeleteInsurances(int? id)
         {
             Insurances obj = db.Insurances.Find(id);
             var STdata = db.SummaryTable.ToList().Where(c => c.STId == obj.STID).Select(c => c).SingleOrDefault();
 
+            var STID = obj.STID;
+            var RelateCash = obj.RelateCash;
             db.Insurances.Remove(obj);
-       
-
             if (obj.PurchaseOrWithdraw == true)
             {
-                var arr = db.CashExpense.Where(C => C.OID == obj.STID).Select(C => C).ToArray();
-                db.CashExpense.Remove(arr[0]);
-                db.SummaryTable.Remove(STdata);
+                if (RelateCash == true)
+                {
+                    var arr = db.CashExpense.Where(C => C.OID == STID).Select(C => C).SingleOrDefault();
+                    db.CashExpense.Remove(arr);
+                }
+                if (RelateCash == false)
+                {
+                    db.SummaryTable.Remove(STdata);
+                }
             }
             else
             {
-                var arr = db.CashIncome.Where(C => C.OID == obj.STID).Select(C => C).ToArray();
-                db.CashIncome.Remove(arr[0]);
-                db.SummaryTable.Remove(STdata);
+                if (RelateCash == true)
+                {
+                    var arr = db.CashIncome.Where(C => C.OID == STID).Select(C => C).SingleOrDefault();
+                    db.CashIncome.Remove(arr);
+                }
+                else
+                {
+                    db.SummaryTable.Remove(STdata);
+                }
             }
-            
+
             db.SaveChanges();
             return RedirectToAction("Index");
         }
